+++ modules/core/04-channel/types/msgs_test.go
@@ -72,12 +72,6 @@ type TypesTestSuite struct {
 	suite.Suite
 
 	proof []byte
-
-	coordinator *ibctesting.Coordinator
-
-	// testing chains used for convenience and readability
-	chainA *ibctesting.TestChain
-	chainB *ibctesting.TestChain
 }
 
 func (suite *TypesTestSuite) SetupTest() {
@@ -107,13 +101,6 @@ func (suite *TypesTestSuite) SetupTest() {
 	suite.Require().NoError(err)
 
 	suite.proof = proof
-}
-
-// SetupTestCoordinator creates a coordinator with 2 test chains.
-func (suite *TypesTestSuite) SetupTestCoordinator() {
-	suite.coordinator = ibctesting.NewCoordinator(suite.T(), 2)
-	suite.chainA = suite.coordinator.GetChain(ibctesting.GetChainID(1))
-	suite.chainB = suite.coordinator.GetChain(ibctesting.GetChainID(2))
 }
 
 func TestTypesTestSuite(t *testing.T) {
