+++ modules/core/04-channel/types/packet_test.go
@@ -51,3 +51,4 @@ func TestPacketValidateBasic(t *testing.T) {
 		}
 	}
 }
+
